+++ DiabloII.Application.Tests/Startup/TestStartup.cs
@@ -1,17 +1,17 @@
 ﻿using System;
 using AutoMapper;
 using DiabloII.Application.Extensions;
+using DiabloII.Application.Tests.Extensions;
 using DiabloII.Domain.Repositories;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
-using DiabloII.Application.Tests.Extensions;
 
-namespace DiabloII.Application.Tests
+namespace DiabloII.Application.Tests.Startup
 {
     public class TestStartup
     {
-        internal static readonly Type ApplicationType = typeof(Startup);
+        internal static readonly Type ApplicationType = typeof(Application.Startup);
         internal static readonly Type DomainType = typeof(IErrorLogRepository);
 
         public void ConfigureServices(IServiceCollection services) => services
